create useUpdateWorks hook
